+++ src/pg_purepy/conversion/basics.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from collections.abc import Callable
-from typing import TYPE_CHECKING, Any, Generic, TypeVar
+from typing import TYPE_CHECKING, Any, override
 
 from pg_purepy.conversion.abc import Converter
 
@@ -9,10 +9,7 @@ if TYPE_CHECKING:
     from pg_purepy.protocol import ConversionContext
 
 
-ConvType = TypeVar("ConvType")
-
-
-class SimpleFunctionConverter(Converter, Generic[ConvType]):
+class SimpleFunctionConverter[ConvType](Converter[ConvType]):
     """
     A simple converter that calls fn_from_pg to convert to a Python type, and fn_to_pg to convert
     from a Python type.
@@ -26,28 +23,32 @@ class SimpleFunctionConverter(Converter, Generic[ConvType]):
         self._from_pg = fn_from_pg
         self._to_pg = fn_to_pg
 
+    @override
     def from_postgres(self, context: ConversionContext, data: str) -> ConvType:
         return self._from_pg(data)
 
+    @override
     def to_postgres(self, context: ConversionContext, data: ConvType) -> str:
         return self._to_pg(data)
 
 
-class BoolConverter(Converter):
+class BoolConverter(Converter[bool]):
     """
     Converter that converts for booleans.
     """
 
     oid = 16
 
-    def from_postgres(self, context: ConversionContext, data: str) -> Any:
+    @override
+    def from_postgres(self, context: ConversionContext, data: str) -> bool:
         return data == "t"
 
+    @override
     def to_postgres(self, context: ConversionContext, data: bool) -> str:
         return "true" if data else "false"
 
 
-class TextConverter(Converter):
+class TextConverter(Converter[str]):
     """
     Text converter for default text types.
     """
@@ -55,20 +56,23 @@ class TextConverter(Converter):
     def __init__(self, oid: int):
         self.oid = oid
 
-    def from_postgres(self, context: ConversionContext, data: str) -> Any:
+    @override
+    def from_postgres(self, context: ConversionContext, data: str) -> str:
         return data
 
-    def to_postgres(self, context: ConversionContext, data: Any) -> str:
+    @override
+    def to_postgres(self, context: ConversionContext, data: str) -> str:
         return str(data)
 
 
-class ByteaConverter(Converter):
+class ByteaConverter(Converter[bytes]):
     """
     Converter that turns bytes objects into bytea objects.
     """
 
     oid = 17
 
+    @override
     def from_postgres(self, context: ConversionContext, data: str) -> Any:
         # bytea data is hex escaped.
         prefix, rest = data[0:2], data[2:]
@@ -76,5 +80,6 @@ class ByteaConverter(Converter):
 
         return bytes.fromhex(rest)
 
+    @override
     def to_postgres(self, context: ConversionContext, data: bytes) -> str:
         return r"\x" + data.hex()