+++ src/pg_purepy/exc.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from collections.abc import Callable
-from typing import TYPE_CHECKING, Self
+from typing import TYPE_CHECKING, Self, override
 
 if TYPE_CHECKING:
     from pg_purepy.connection import AsyncPostgresConnection
@@ -21,6 +21,7 @@ class ConnectionForciblyKilledError(PostgresqlError):
     def __init__(self, conn: AsyncPostgresConnection):
         self._conn = conn
 
+    @override
     def __str__(self) -> str:
         return f"Connection {self._conn!r} could not send the Terminate message"